# docs/extra/openapi-extra-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "extra API",
    "version": "1.0.0",
    "description": "The exm.co.il public API. Each section below is an API namespace with its own availability notes and conventions.\n\n## Authentication\n\nEvery operation requires a Bearer token - your account API key from [/my/api/](https://www.exm.co.il/my/api/):\n\n```http\nAuthorization: Bearer YOUR_API_KEY\n```\n\nA missing or invalid key returns **401** with a JSON error body (the exact envelope varies per namespace and is documented on each operation). The key grants access to the whole account - keep it server-side only; never embed it in browser or app code.\n\n## Conventions\n\n- ids are opaque string identifiers\n- times are ISO 8601 with timezone offset (Asia/Jerusalem local), e.g. `2026-07-12T15:00:00+03:00`\n- phone numbers are E.164; where a field omits the `+` (e.g. `972501234567`) the operation says so\n"
  },
  "servers": [
    {
      "url": "https://api.exm.co.il/v1",
      "description": "Production"
    },
    {
      "url": "https://www.exm.co.il/api/v1",
      "description": "Production (main-site base, same API)"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/calls/": {
      "post": {
        "tags": [
          "Calls"
        ],
        "summary": "Get calls history (filtering + pagination)",
        "operationId": "getCallsHistory",
        "description": "The account's call history, newest first (`start_time DESC, id DESC`),\nwith filtering and opaque-cursor pagination.\n\n**Pagination**: the response `metadata.next_page_token` is an opaque\ncursor bound to the exact filter that produced it (the filter is\nbound to the token). To fetch the next page, repeat the SAME\nrequest and add the token as `pagination.next`. Changing any filter\nwhile reusing an old token fails with error code `5010`\n(\"Invalid next token provided: hash does not match.\"). A `null`\n`next_page_token` means there are no more pages.\n\n**Errors** (HTTP 200, `success: false`, `errors` array):\n\n| code | id | Meaning |\n|---|---|---|\n| 5010 | next_token | `pagination.next` token does not match the current filter |\n",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "A single call's id - fetch just that call.",
                    "example": "zLvdabAqcT5"
                  },
                  "freetext": {
                    "type": "string",
                    "description": "Free-text search over the calls.",
                    "example": "0501234567"
                  },
                  "starred": {
                    "type": "boolean",
                    "description": "When truthy, only starred calls."
                  },
                  "call_types": {
                    "type": "array",
                    "description": "Call type aliases (any of): `incoming`, `incoming_missed`,\n`incoming_start`, `outgoing`, `outgoing_missed`,\n`outgoing_start`. Unknown aliases are ignored.\n",
                    "items": {
                      "type": "string",
                      "enum": [
                        "incoming",
                        "incoming_missed",
                        "incoming_start",
                        "outgoing",
                        "outgoing_missed",
                        "outgoing_start"
                      ]
                    },
                    "example": [
                      "incoming",
                      "incoming_missed"
                    ]
                  },
                  "line_types": {
                    "type": "array",
                    "description": "Service (line) type enums, any of: `1` = virtual number,\n`2` = virtual IVR, `5` = mobile line, `7` = SIP line.\n",
                    "items": {
                      "type": "integer",
                      "enum": [
                        1,
                        2,
                        5,
                        7
                      ]
                    },
                    "example": [
                      1,
                      2
                    ]
                  },
                  "direction": {
                    "type": "string",
                    "deprecated": true,
                    "description": "DO NOT USE. Accepted by the server but currently\nmalfunctioning: it overwrites the `line_types` filter with\ncall-type values instead of filtering by direction, so the\nresults are wrong. Use `call_types` to filter by direction\n(e.g. `[\"incoming\", \"incoming_missed\"]`).\n"
                  },
                  "duration": {
                    "type": "object",
                    "description": "Call duration bounds, in seconds (inclusive).",
                    "properties": {
                      "gte": {
                        "type": "integer",
                        "description": "Minimum duration (seconds).",
                        "example": 30
                      },
                      "lte": {
                        "type": "integer",
                        "description": "Maximum duration (seconds).",
                        "example": 600
                      }
                    }
                  },
                  "time": {
                    "type": "object",
                    "description": "Call start-time bounds. Each bound is any strtotime()-parseable\nstring (e.g. `2026-07-01`, `2026-07-01 08:00:00`); unparseable\nvalues are silently ignored.\n",
                    "properties": {
                      "from": {
                        "type": "string",
                        "example": "2026-07-01"
                      },
                      "to": {
                        "type": "string",
                        "example": "2026-07-31 23:59:59"
                      }
                    }
                  },
                  "pagination": {
                    "type": "object",
                    "properties": {
                      "items": {
                        "type": "integer",
                        "minimum": 20,
                        "maximum": 100,
                        "default": 20,
                        "description": "Items per page (clamped to 20..100)."
                      },
                      "next": {
                        "type": "string",
                        "description": "Opaque cursor from the previous response's metadata.next_page_token. Must be sent with the exact same filter.\n"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The matching calls page - or, on a handler-level error (e.g. bad pagination token), an error body with success:false.\n",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "title": "Calls page",
                      "properties": {
                        "metadata": {
                          "type": "object",
                          "properties": {
                            "items_per_page": {
                              "type": "integer",
                              "example": 20
                            },
                            "response_count": {
                              "type": "integer",
                              "example": 20,
                              "description": "Number of calls in this response."
                            },
                            "next_page_token": {
                              "type": "string",
                              "nullable": true,
                              "description": "Cursor for the next page; null when this is the last page."
                            }
                          }
                        },
                        "calls": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Call"
                          }
                        }
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ErrorsArrayEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/CallsUnauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ApiBlocked"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Standard Bearer token auth (`Authorization: Bearer <token>`). Use your account's API key from /my/api/ (TokensModel TYPE_API). The same key authenticates every exm.co.il API endpoint."
      }
    },
    "responses": {
...
      "CallsUnauthorized": {
        "description": "Missing/invalid/expired Bearer token, or blocked account",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthError"
            },
            "example": {
              "success": false,
              "errors": {
                "id": "HTTP_401",
                "code": 401,
                "description": "401 Unauthorized"
              }
            }
          }
        }
      },
      "ApiBlocked": {
        "description": "The account's API access is administratively blocked",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthError"
            },
            "example": {
              "success": false,
              "errors": {
                "id": "HTTP_403",
                "code": 403,
                "description": "403 Forbidden"
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "AuthError": {
        "type": "object",
        "description": "Auth-layer error envelope (`errors` is a single object here).",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "errors": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "HTTP_401"
              },
              "code": {
                "type": "integer",
                "example": 401
              },
              "description": {
                "type": "string",
                "example": "401 Unauthorized"
              }
            }
          }
        }
      },
...
      "ErrorsArrayEnvelope": {
        "type": "object",
        "description": "Handler-level error envelope used by POST /calls/ and GET /calls/recording/ - `errors` is an ARRAY of error objects. Returned with HTTP status 200.\n",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "Machine-readable error key (usually the offending field).",
                  "example": "call_id"
                },
                "code": {
                  "type": "integer",
                  "description": "Numeric error code (see the operation's error table).",
                  "example": 7020
                },
                "description": {
                  "type": "string",
                  "example": "no call found"
                }
              }
            }
          }
        }
      },
      "Call": {
        "type": "object",
        "description": "One call, as returned by the calls list. This is the same shape as\nthe automations webhook POST payload, plus API-only enrichments\n(`numbers.own.color`, `numbers.own.line_type`, `contact.avatar`).\n\nThe `contact`, `recording` and `AI` keys are `false` (boolean) when\nabsent, an object when present.\n",
        "properties": {
          "id": {
            "type": "string",
            "description": "The call's id.",
            "example": "zLvdabAqcT5"
          },
          "event": {
            "type": "string",
            "enum": [
              "start",
              "end"
            ],
            "description": "Lifecycle stage - completed calls are `end`, in-progress calls are `start`."
          },
          "type": {
            "type": "string",
            "enum": [
              "incoming",
              "incoming_missed",
              "incoming_start",
              "outgoing",
              "outgoing_missed",
              "outgoing_start"
            ],
            "description": "Call type alias.",
            "example": "incoming"
          },
          "own_type": {
            "type": "string",
            "enum": [
              "VNUM",
              "VIVR",
              "MOBL",
              "SIPL"
            ],
            "description": "What kind of line \"own\" is: VNUM = virtual number, VIVR = virtual IVR, MOBL = mobile line, SIPL = SIP line.\n"
          },
          "ivr_dialed": {
            "type": "string",
            "nullable": true,
            "description": "The IVR menu digit the caller dialed (null when not an IVR call).",
            "example": "2"
          },
          "duration": {
            "type": "integer",
            "nullable": true,
            "description": "Call duration in seconds (null while the call is in progress).",
            "example": 74
          },
          "note": {
            "type": "string",
            "nullable": true,
            "description": "The user's note on the call."
          },
          "starred": {
            "type": "boolean",
            "description": "Whether the user starred the call."
          },
          "time": {
            "type": "object",
            "properties": {
              "timezone": {
                "type": "string",
                "example": "Asia/Jerusalem"
              },
              "gmt_offset": {
                "type": "string",
                "example": "+03:00"
              },
              "start": {
                "type": "string",
                "description": "Local start time, `Y-m-d H:i:s` (in `timezone`).",
                "example": "2026-07-09 14:32:05"
              },
              "end": {
                "type": "string",
                "nullable": true,
                "description": "Local end time, `Y-m-d H:i:s` (absent/null while in progress).",
                "example": "2026-07-09 14:33:19"
              }
            }
          },
          "numbers": {
            "type": "object",
            "properties": {
              "own": {
                "type": "object",
                "description": "The account's own line the call went through.",
                "properties": {
                  "friendly": {
                    "type": "string",
                    "example": "050-1234567"
                  },
                  "e164": {
                    "type": "string",
                    "example": "+972501234567"
                  },
                  "color": {
                    "type": "string",
                    "nullable": true,
                    "description": "The user-assigned line color (API/app-only enrichment; null when unset).",
                    "example": "#ce081e"
                  },
                  "line_type": {
                    "type": "string",
                    "nullable": true,
                    "enum": [
                      "mobile",
                      "vn",
                      "ivr",
                      "sip",
                      null
                    ],
                    "description": "API/app-only enrichment - which kind of line the call came through."
                  }
                }
              },
              "caller": {
                "type": "object",
                "description": "The call's initiator. For withheld caller id both fields hold the literal string `ANONYMOUS`.\n",
                "properties": {
                  "friendly": {
                    "type": "string",
                    "example": "050-1234567"
                  },
                  "e164": {
                    "type": "string",
                    "example": "+972501234567"
                  }
                }
              },
              "destination": {
                "type": "object",
                "description": "The call's receiver.",
                "properties": {
                  "friendly": {
                    "type": "string",
                    "nullable": true,
                    "example": "050-1234567"
                  },
                  "e164": {
                    "type": "string",
                    "nullable": true,
                    "example": "+972501234567"
                  }
                }
              }
            }
          },
          "contact": {
            "description": "The matched contact, or boolean `false` when the other party is not in the user's contacts.\n",
            "nullable": true,
            "properties": {
              "id": {
                "type": "string",
                "description": "Contact id.",
                "example": "c2yFfPaRcT5"
              },
              "book_id": {
                "type": "string",
                "description": "Contact-book id.",
                "example": "Jx8ObCAucT5"
              },
              "name": {
                "type": "string",
                "example": "ישראל ישראלי"
              },
              "avatar": {
                "type": "string",
                "nullable": true,
                "description": "Contact photo URL (API/app-only enrichment)."
              }
            }
          },
          "recording": {
            "description": "Recording marker: an empty object `{}` when the call has a recording, boolean `false` when it doesn't. To fetch the actual file use GET /calls/recording/ or POST /calls/get-recording-urls/.\n",
            "nullable": true
          },
          "AI": {
            "description": "The GenieAI call summary, or boolean `false` when none exists.\n",
            "nullable": true,
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "PENDING",
                  "COMPLETED"
                ]
              },
              "summary": {
                "type": "object",
                "properties": {
                  "gist": {
                    "type": "string",
                    "description": "One-line summary title.",
                    "example": "בירור לגבי התאמת השירות"
                  },
                  "body": {
                    "type": "string",
                    "description": "The full summary text."
                  },
                  "next_steps": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Actionable next-step items the AI extracted from the call."
                  }
                }
              }
            }
          },
          "recording_excluded": {
            "type": "boolean",
            "description": "Present only on accounts enrolled in the recording-exclusions feature - true when the recording was deliberately skipped by the line's exclusion list (distinguishes \"excluded\" from plain \"no recording\").\n"
          }
        }
      },
...
    }
  },
...
}